requirements.txt
aiogram>=2.25.1
aiohttp>=3.8.5
python-dotenv>=1.0.0
SQLAlchemy>=2.0.19
aiosqlite>=0.19.0
asyncio>=3.4.3
cryptography>=41.0.3
alembic>=1.12.0
loguru>=0.7.0
aiofiles>=23.1.0
python-dateutil>=2.8.2 